requirements-build.txt
# These are the extra requirements required to build but not for the lambda itself
# ex: set ft=sh:

pyyaml
setuptools

black>=20.8b1
flake8>=3.8.4
flake8-black>=0.2.1
flake8-builtins>=1.5.3
flake8-docstrings>=1.5.0
flake8-import-order>=0.18.1
flake8-return>=1.1.2
flake8-rst-docstrings>=0.0.14
pep8-naming>=0.11.1

yamllint>=1.26.0

ruff
pytest
pytest-xdist
pytest-cov
twine

mkdocs
mkdocs-material
mkdocs-material-extensions
mkdocstrings[python]
